moved .py to fs42 and update .sh
