+++ rebuild_schedule.sh
@@ -4,4 +4,4 @@
 cd $(dirname $0)
 . env/bin/activate
 
-python3 station_42.py
+python3 fs42/station_42.py
